+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+* Send JSON responses from controller
+
 ## [3.0.0-beta] - 2016-01-23
 ### Fixed
 * Fix typo preventing csv harvesting